9:50 sunday only one marker viable map fragment Ready search gm places Left Visit : https://developers.google.com/maps/documentation/places/web-service/search-nearby
And :
https://developers.google.com/maps/documentation/places/web-service/search

+++ app/src/main/java/com/example/finalproject1_11/Model/DB/LocationTable.kt
@@ -9,8 +9,8 @@ data class LocationTable(
     @PrimaryKey(autoGenerate = true)
     var id: Int,
     var name : String,
-    var note : String,
-    var longitude : Long,
-    var latitude : Long,
-    var whatKind : String
+    var note : String, // why visiting
+    var longitude : Float,
+    var latitude : Float,
+    var whatKind : String // is it a restaurant or private
     )
